+++ character/tests/conftest.py
@@ -17,6 +17,7 @@ def django_db_setup(django_db_setup, django_db_blocker):
         call_command(loaddata.Command(), "character/fixtures/species_traits.yaml")
         call_command(loaddata.Command(), "character/fixtures/species.yaml")
         call_command(loaddata.Command(), "character/fixtures/feats.yaml")
+        call_command(loaddata.Command(), "character/fixtures/magic_items.yaml")
 
 
 @pytest.fixture(scope="function")